rename namespace from latin america to south america
